docker/heic-to-jpg/requirements.txt
boto3
pillow
pillow-heif
